+++ src/main/java/com/neovisionaries/jdo/TaskExecutor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2017 Neo Visionaries Inc.
+ * Copyright (C) 2014-2022 Neo Visionaries Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -59,7 +59,8 @@ import javax.jdo.Transaction;
  */
 public class TaskExecutor
 {
-    private final PersistenceManagerFactory factory;
+    private PersistenceManagerFactory mFactory;
+    private PersistenceManager mManager;
 
 
     /**
@@ -80,7 +81,31 @@ public class TaskExecutor
             throw new IllegalArgumentException("factory is null.");
         }
 
-        this.factory = factory;
+        this.mFactory = factory;
+    }
+
+
+    /**
+     * Constructor with a persistence manager.
+     *
+     * @param manager
+     *         A persistence manager. Must not be {@code null}.
+     *
+     * @throws IllegalArgumentException
+     *         {@code manager} is {@code null}.
+     *
+     * @since 1.19
+     */
+    public TaskExecutor(PersistenceManager manager)
+    {
+        // Check if the manager is null.
+        if (manager == null)
+        {
+            // Manager must not be null.
+            throw new IllegalArgumentException("manager is null.");
+        }
+
+        this.mManager = manager;
     }
 
 
@@ -148,8 +173,71 @@ public class TaskExecutor
     /**
      * Execute a task.
      *
+     * <p>
+     * This method is an alias of {@link #execute(Task, boolean, int)
+     * execute(task, false, retryCount)}.
+     * </p>
+     *
      * @param task
      *         Task to be executed.
+     *
+     * @param retryCount
+     *         Indicates how many times to retry the task. The task is retried
+     *         only when {@code task.}{@link Task#run(PersistenceManager)
+     *         run(PersistenceManager)} threw a {@link JDOCanRetryException}.
+     *         If an exception of other type was thrown, the task is not retried.
+     *
+     * @return
+     *         The object returned from {@code task.}{@link Task#run(PersistenceManager)
+     *         run(PersistenceManager)}.
+     *
+     * @throws IllegalArgumentException
+     *         {@code task} is {@code null}, or {@code retryCount} is less than 0.
+     *
+     * @throws JDOException
+     *         A persistence manager could not be obtained or the task failed.
+     *         See the description of {@link #execute(Task, boolean, int)}
+     *         for details.
+     *
+     * @since 1.19
+     */
+    public Object execute(Task task, int retryCount)
+    {
+        return execute(task, false, retryCount);
+    }
+
+
+    /**
+     * Execute a task.
+     *
+     * <p>
+     * If this {@code TaskExecutor} instance was created by the constructor
+     * {@link #TaskExecutor(PersistenceManagerFactory)}, a new instance of
+     * {@link PersistenceManager} is created inside this {@code execute} method
+     * by calling the {@link PersistenceManagerFactory#getPersistenceManager()
+     * getPersistenceManager()} method of the {@link PersistenceManagerFactory}
+     * instance that was passed via the constructor.
+     * </p>
+     *
+     * <p>
+     * On the other hand, if this {@code TaskExecutor} instance was created by
+     * the constructor {@link #TaskExecutor(PersistenceManager)}, the
+     * {@link PersistanceManager} instance passed via the constructor is used
+     * inside this {@code execute} method.
+     * </p>
+     *
+     * <p>
+     * Note that if the current transaction of the {@link PersistenceManager}
+     * instance that was passed via the constructor
+     * {@link #TaskExecutor(PersistenceManager)} is already active, transaction
+     * operations are NOT executed even if the {@code transaction} argument is
+     * {@code true}.
+     * </p>
+     *
+     * @param task
+     *         Task to be executed. If the given instance implements the
+     *         {@link TransactionAwareTask} interface, transaction-related
+     *         callback methods of the interface will be called appropriately.
      *
      * @param transaction
      *         {@code true} to execute the task in a transaction.
@@ -198,12 +286,75 @@ public class TaskExecutor
             throw new IllegalArgumentException("retryCount < 0");
         }
 
+        // If this TaskExecutor instance was created by TaskExecutor(PersistenceManager),
+        // the PersistenceManager instance given via the constructor is used.
+        //
+        // Otherwise, a new PersistenceManager instance is created here by calling the
+        // getPersistenceManager() method of the PersistenceManagerFactory instance
+        // that was given via TaskExecutor(PersistenceManagerFactory).
+        PersistenceManager manager = (mManager != null) ? mManager : createManager();
+        boolean managerCreated     = (mManager == null);
+
+        try
+        {
+            // Execute the task.
+            return doExecute(task, transaction, retryCount, manager);
+        }
+        finally
+        {
+            if (managerCreated)
+            {
+                // Close the persistence manager in any case.
+                manager.close();
+            }
+        }
+    }
+
+
+    /**
+     * Get the persistence manager factory that was given to the
+     * {@link #TaskExecutor(PersistenceManagerFactory)} constructor.
+     *
+     * @return
+     *         The instance of {@link PersistenceManagerFactory} that was given
+     *         to the {@link #TaskExecutor(PersistenceManagerFactory)} constructor.
+     *         {@code null} is returned if this {@code TaskExecutor} instance was
+     *         created by other constructors.
+     *
+     * @since 1.19
+     */
+    public PersistenceManagerFactory getFactory()
+    {
+        return mFactory;
+    }
+
+
+    /**
+     * Get the persistence manager that was given to the
+     * {@link #TaskExecutor(PersistenceManager)} constructor.
+     *
+     * @return
+     *         The instance of {@link PersistenceManager} that was given
+     *         to the {@link #TaskExecutor(PersistenceManager)} constructor.
+     *         {@code null} is returned if this {@code TaskExecutor} instance was
+     *         created by other constructors.
+     *
+     * @since 1.19
+     */
+    public PersistenceManager getManager()
+    {
+        return mManager;
+    }
+
+
+    private PersistenceManager createManager()
+    {
         PersistenceManager manager = null;
 
         try
         {
             // Get a persistence manager from the persistence manager factory.
-            manager = factory.getPersistenceManager();
+            manager = mFactory.getPersistenceManager();
         }
         catch (JDOException e)
         {
@@ -222,28 +373,25 @@ public class TaskExecutor
             throw new JDOException("getPersistenceManager() returned null.");
         }
 
-        try
-        {
-            // Execute the task.
-            return doExecute(task, transaction, retryCount, manager);
-        }
-        finally
-        {
-            // Close the persistence manager in any case.
-            manager.close();
-        }
+        return manager;
     }
 
 
     private Object doExecute(Task task, boolean transaction, int retryCount, PersistenceManager manager)
     {
-        Transaction tx = null;
+        Transaction tx = manager.currentTransaction();
+
+        // A flag that indicates whether to do transaction operations.
+        // If the current transaction is already active, transaction
+        // operations are not executed even if the 'transaction' flag
+        // passed to the 'execute' method is true.
+        boolean doTx = transaction && !tx.isActive();
 
         // If transaction is required.
-        if (transaction)
+        if (doTx)
         {
             // Begin a transaction.
-            tx = beginTransaction(manager, task);
+            beginTransaction(manager, task, tx);
         }
 
         try
@@ -251,7 +399,7 @@ public class TaskExecutor
             // Execute the task
             Object ret = doExecute(task, retryCount, manager);
 
-            if (tx != null)
+            if (doTx)
             {
                 // Commit the changes.
                 commitTransaction(manager, task, tx);
@@ -262,7 +410,7 @@ public class TaskExecutor
         finally
         {
             // If transaction is required but not committed.
-            if (tx != null && tx.isActive())
+            if (doTx && tx.isActive())
             {
                 // Revoke the changes.
                 rollbackTransaction(manager, task, tx);
@@ -284,11 +432,8 @@ public class TaskExecutor
     }
 
 
-    private static Transaction beginTransaction(PersistenceManager manager, Task task)
+    private static void beginTransaction(PersistenceManager manager, Task task, Transaction tx)
     {
-        // Get the current transaction.
-        Transaction tx = manager.currentTransaction();
-
         // Cast by TransactionAwareTask if possible.
         TransactionAwareTask tat = cast(task);
 
@@ -307,8 +452,6 @@ public class TaskExecutor
             // Call a hook after beginning a transaction.
             tat.afterTransactionBegin(manager, tx);
         }
-
-        return tx;
     }